+++ spec/osm/flexi_record_spec.rb
@@ -85,6 +85,67 @@ describe "Flexi Record" do
       fields[3].id.should == 'f_2'
     end
 
+    it "Add field (success)" do
+      url = "https://www.onlinescoutmanager.co.uk/extras.php?action=addColumn&sectionid=1&extraid=2"
+
+      post_data = {
+        'apiid' => @CONFIGURATION[:api][:osm][:id],
+        'token' => @CONFIGURATION[:api][:osm][:token],
+        'userid' => 'user_id',
+        'secret' => 'secret',
+        'columnName' => 'name',
+      }
+
+      data = {
+        "extraid" => "2",
+        "sectionid" => "1",
+        "name" => "A Flexi Record",
+        "config" => "[{\"id\":\"f_1\",\"name\":\"name\",\"width\":\"150\"}]",
+        "total" => "none",
+        "extrafields" => "[]",
+        "structure" => [
+          {
+            "rows" => [
+              {"name" => "First name","field" => "firstname","width" => "150px"},
+              {"name" => "Last name","field" => "lastname","width" => "150px"},
+            ],
+            "noscroll" => true
+          },
+          {"rows" => [
+            {"name" => "name","field" => "f_1","width" => "150px","editable" => true},
+          ]}
+        ]
+      }
+      HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>data.to_json}) }
+
+      Osm::FlexiRecord.add_field(@api, 1, 2, 'name').should be_true
+    end
+
+    it "Add field (failure)" do
+      data = {
+        "extraid" => "2",
+        "sectionid" => "1",
+        "name" => "A Flexi Record",
+        "config" => "[]",
+        "total" => "none",
+        "extrafields" => "[]",
+        "structure" => [
+          {
+            "rows" => [
+              {"name" => "First name","field" => "firstname","width" => "150px"},
+              {"name" => "Last name","field" => "lastname","width" => "150px"},
+            ],
+            "noscroll" => true
+          },
+          {"rows" => [
+          ]}
+        ]
+      }
+      HTTParty.should_receive(:post) { DummyHttpResult.new(:response=>{:code=>'200', :body=>data.to_json}) }
+
+      Osm::FlexiRecord.add_field(@api, 1, 2, 'name').should be_false
+    end
+
     it "Fetch Data" do
       data = {
         'identifier' => 'scoutid',